Read all

+++ int/PerformerData.php
@@ -1,6 +1,6 @@
 <?php
   include_once("fest.php");
-  A_Check('Committee','Dance');
+  A_Check('Staff');
 
   dostaffhead("Performer Data");
   include_once("DanceLib.php");
@@ -14,10 +14,13 @@
   $id = $_REQUEST['id'];
   $Side = Get_Side($id);
 
+  $UpdateValid = ($Side['IsASide'] && Access('Staff','Dance') || ($Side['IsAnAct'] && Access('Staff','Music')) || ($Side['IsOther'] && Access('Staff','Other')));
+  
   echo '<h2>Performer Files For ' . $Side['SN'] . '</h2>';
   switch ($_REQUEST['ACTION']) {
   
   case 'STORE':
+    if (!$UpdateValid) break;
     $target_dir = "Store/Performers/$id";
     
     mkdir($target_dir,0777,true);
@@ -43,6 +46,7 @@
     break;
     
   case 'DELETE':
+    if (!$UpdateValid) break;
     $fname = base64_decode($_REQUEST['f']);
     if (file_exists("Store/Performers/$id/$fname")) {
       if (unlink("Store/Performers/$id/$fname")) {
@@ -71,7 +75,8 @@
         $fname = basename($file);
         echo "<tr><td><a href=ShowFile.php?l64=" . base64_encode("Store/Performers/$id/$fname") . ">$fname</a>";
         echo "<td>" . formatBytes(filesize($file)) . "<td>" . date('j/m/Y H:i:s',filectime($file));
-        echo "<td><a href=ShowFile.php?D=Store/Performers/$id/$fname>download</a>, <a href=PerformerData.php?id=$id&ACTION=DELETE&f=" . base64_encode($fname) . ">delete</a>";
+        echo "<td><a href=ShowFile.php?D=Store/Performers/$id/$fname>download</a>";
+        if ($UpdateValid) echo ", <a href=PerformerData.php?id=$id&ACTION=DELETE&f=" . base64_encode($fname) . ">delete</a>";
       }
       echo "</table><p>";
     } else {
@@ -80,14 +85,15 @@
 
   
   // Upload && back to edit performer
-  echo '<form action="PerformerData.php" method="post" enctype="multipart/form-data" id=Uploads>';
-    echo "Select file(s) to upload:";
-    echo fm_hidden('id', $id);
-    echo fm_hidden('ACTION', 'STORE');
-    echo '<input type="file" name="uploads[]" multiple onchange=this.form.submit()>';
-    echo " &nbsp; &nbsp; Do not upload more than 15M at once, for large files contact <a href=mailto:Richard@wavwebs.com>Richard</a>.\n";
-    echo "</form>\n";
-
+  if ($UpdateValid) {
+    echo '<form action="PerformerData.php" method="post" enctype="multipart/form-data" id=Uploads>';
+      echo "Select file(s) to upload:";
+      echo fm_hidden('id', $id);
+      echo fm_hidden('ACTION', 'STORE');
+      echo '<input type="file" name="uploads[]" multiple onchange=this.form.submit()>';
+      echo " &nbsp; &nbsp; Do not upload more than 15M at once, for large files contact <a href=mailto:Richard@wavwebs.com>Richard</a>.\n";
+      echo "</form>\n";
+  };
       
   echo "<h2><a href=" . ($Side['IsASide']?'AddDance.php':'AddMusic.php') . "?sidenum=$id>Back to " . $Side['SN'] . "</a></h2>\n";
